Changes to Catalog

+++ src/Catalog/Catalog.Contracts/Events.cs
@@ -2,7 +2,7 @@ namespace YourBrand.Catalog.Contracts;
 
 public sealed record ProductDetailsUpdated
 {
-    public required long ProductId { get; init; }
+    public required int ProductId { get; init; }
 
     public required string Name { get; init; }
 
@@ -11,7 +11,7 @@ public sealed record ProductDetailsUpdated
 
 public sealed record ProductPriceUpdated
 {
-    public required long ProductId { get; init; }
+    public required int ProductId { get; init; }
 
     public required decimal NewPrice { get; init; }
 
@@ -22,7 +22,7 @@ public sealed record ProductPriceUpdated
 
 public sealed record ProductVatRateUpdated
 {
-    public required long ProductId { get; init; }
+    public required int ProductId { get; init; }
 
     public required double? NewVatRate { get; init; }
 }
@@ -30,21 +30,21 @@ public sealed record ProductVatRateUpdated
 
 public sealed record ProductImageUpdated
 {
-    public required long ProductId { get; init; }
+    public required int ProductId { get; init; }
 
     public required string ImageUrl { get; init; }
 }
 
 public sealed record ProductHandleUpdated
 {
-    public required long ProductId { get; init; }
+    public required int ProductId { get; init; }
 
     public required string Handle { get; init; }
 }
 
 public sealed record ProductListingStateUpdated
 {
-    public required long ProductId { get; init; }
+    public required int ProductId { get; init; }
 
     public required ProductListingState ListingState { get; init; }
 }
@@ -57,7 +57,7 @@ public enum ProductListingState
 
 public sealed record ProductSkuUpdated
 {
-    public required long ProductId { get; init; }
+    public required int ProductId { get; init; }
 
     public required string Sku { get; init; }
 }
